Finish Api/User

+++ pdv/src/main/java/com/api/pdv/controller/UserController.java
@@ -4,6 +4,7 @@ import com.api.pdv.middleware.UserMiddleware;
 import com.api.pdv.model.User;
 import com.api.pdv.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -18,6 +19,8 @@ public class UserController {
 
     @Autowired
     private UserService userService;
+
+    private final HttpHeaders responseHeaders = new HttpHeaders();
 
     @PostMapping(
             value = "/cadastrar",
@@ -46,45 +49,89 @@ public class UserController {
         }
     }
 
-    @GetMapping( value = "/list")
-    public ResponseEntity<List<User>> getUsers() {
+    @PostMapping(
+            value = "/login",
+            consumes = MediaType.APPLICATION_JSON_VALUE,
+            produces = MediaType.APPLICATION_JSON_VALUE
+    )
+    public ResponseEntity<String> loginUser(@RequestBody User user) {
+        String md5 = UserMiddleware.getHashMd5(user.getPassword());
+        String encoded = UserMiddleware.getEncode64(md5);
+        user.setPassword(encoded);
+        User u = this.userService.loginService(user.getName(), user.getPassword());
         try {
-            List<User> users = this.userService.getUsers();
-            return ResponseEntity.status(HttpStatus.ACCEPTED).body(users);
+            if (u != null) {
+                UUID id = u.getId();
+                responseHeaders.set(UserMiddleware.getIdInfo(), String.valueOf(id));
+                return ResponseEntity.status(HttpStatus.ACCEPTED).body(String.valueOf(id));
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("User/Not/Exist");
+            }
         } catch (Exception e) {
             return ResponseEntity.internalServerError().build();
         }
     }
 
-    @GetMapping( value = "/{id}")
-    public ResponseEntity<User> getUser(@PathVariable UUID id) {
-        User u = this.userService.findUserById(id);
-        try {
-            if (u == null) {
-                return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body(null);
+    @GetMapping( value = "/list")
+    public ResponseEntity<?> getUsers() {
+        List<String> result = responseHeaders.get(UserMiddleware.getIdInfo());
+        if (result == null || result.isEmpty()) {
+            return ResponseEntity.badRequest().build();
+        } else {
+            try {
+                User user = this.userService.findUserById(UUID.fromString(result.stream().findFirst().get()));
+                if (user == null) {
+                    return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("User/Not/Exist");
+                }
+                List<User> users = this.userService.getUsers();
+                return ResponseEntity.status(HttpStatus.ACCEPTED).body(users);
+            } catch (Exception e) {
+                return ResponseEntity.internalServerError().build();
             }
-            return ResponseEntity.status(HttpStatus.ACCEPTED).body(u);
-        } catch (Exception e) {
-            return ResponseEntity.internalServerError().build();
+        }
+    }
+
+    @GetMapping( value = "/{id}")
+    public ResponseEntity<?> getUser(@PathVariable UUID id) {
+        List<String> result = responseHeaders.get(UserMiddleware.getIdInfo());
+        if (result == null || result.isEmpty()) {
+            return ResponseEntity.badRequest().build();
+        } else {
+            try {
+                User u = this.userService.findUserById(id);
+                User user = this.userService.findUserById(UUID.fromString(result.stream().findFirst().get()));
+                if (u == null || user == null) {
+                    return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("User/Not/Exist");
+                }
+                return ResponseEntity.status(HttpStatus.ACCEPTED).body(u);
+            } catch (Exception e) {
+                return ResponseEntity.internalServerError().build();
+            }
         }
     }
 
     @PutMapping( value = "/update/{id}")
     public ResponseEntity<?> updateUser(@PathVariable UUID id, @RequestBody User user) {
-        User u = this.userService.findUserById(id);
-        if (u == null) {
-            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("User/Not/Exist");
+        List<String> result = responseHeaders.get(UserMiddleware.getIdInfo());
+        if (result == null || result.isEmpty()) {
+            return ResponseEntity.badRequest().build();
         } else {
             try {
-                Boolean isTrue = UserMiddleware.ValidUser(user.getName(), user.getPassword().trim());
-                if (isTrue) {
-                    String md5 = UserMiddleware.getHashMd5(user.getPassword());
-                    String encoded = UserMiddleware.getEncode64(md5);
-                    user.setPassword(encoded);
-                    this.userService.updateUser(user, id);
-                    return ResponseEntity.status(HttpStatus.ACCEPTED).body(u);
+                User u = this.userService.findUserById(id);
+                User us = this.userService.findUserById(UUID.fromString(result.stream().findFirst().get()));
+                if (u == null || us == null) {
+                    return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("User/Not/Exist");
                 } else {
-                    return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("Inavalid/Camps");
+                    Boolean isTrue = UserMiddleware.ValidUser(user.getName(), user.getPassword().trim());
+                    if (isTrue) {
+                        String md5 = UserMiddleware.getHashMd5(user.getPassword());
+                        String encoded = UserMiddleware.getEncode64(md5);
+                        user.setPassword(encoded);
+                        this.userService.updateUser(user, id);
+                        return ResponseEntity.status(HttpStatus.ACCEPTED).body(u);
+                    } else {
+                        return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("Inavalid/Camps");
+                    }
                 }
             } catch (Exception e) {
                 return ResponseEntity.internalServerError().body(e.getMessage());
@@ -94,26 +141,41 @@ public class UserController {
 
     @DeleteMapping( value = "/delete/{id}" )
     public ResponseEntity<String> deleteUserById(@PathVariable UUID id) {
-        User user = this.userService.findUserById(id);
-        try {
-            if (user == null) {
-                return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("User/Not/Exist");
-            } else {
-                this.userService.deleteUser(id);
-                return ResponseEntity.status(HttpStatus.ACCEPTED).body("User/Deleted");
+        List<String> result = responseHeaders.get(UserMiddleware.getIdInfo());
+        if (result == null || result.isEmpty()) {
+            return ResponseEntity.badRequest().build();
+        } else {
+            try {
+                User user = this.userService.findUserById(id);
+                User u = this.userService.findUserById(UUID.fromString(result.stream().findFirst().get()));
+                if (user == null || u == null) {
+                    return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("User/Not/Exist");
+                } else {
+                    this.userService.deleteUser(id);
+                    return ResponseEntity.status(HttpStatus.ACCEPTED).body("User/Deleted");
+                }
+            } catch (Exception e) {
+                return ResponseEntity.internalServerError().build();
             }
-        } catch (Exception e) {
-            return ResponseEntity.internalServerError().build();
         }
     }
 
     @DeleteMapping( value = "/delete/all" )
     public ResponseEntity<String> deleteAllUsers() {
-        try {
-            this.userService.deleteAllUsers();
-            return ResponseEntity.status(HttpStatus.ACCEPTED).body("Users/Deleted");
-        } catch (Exception e) {
-            return ResponseEntity.internalServerError().build();
+        List<String> result = responseHeaders.get(UserMiddleware.getIdInfo());
+        if (result == null || result.isEmpty()) {
+            return ResponseEntity.badRequest().build();
+        } else {
+            try {
+                User user = this.userService.findUserById(UUID.fromString(result.stream().findFirst().get()));
+                if (user == null) {
+                    return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("User/Not/Exist");
+                }
+                this.userService.deleteAllUsers();
+                return ResponseEntity.status(HttpStatus.ACCEPTED).body("Users/Deleted");
+            } catch (Exception e) {
+                return ResponseEntity.internalServerError().build();
+            }
         }
     }
